+++ runner/monitor/java/androidx/test/platform/io/PlatformTestStorageRegistry.java
@@ -17,6 +17,7 @@ package androidx.test.platform.io;
 
 import static androidx.test.internal.util.Checks.checkNotNull;
 
+import android.net.Uri;
 import androidx.annotation.NonNull;
 import androidx.test.annotation.ExperimentalTestApi;
 import androidx.test.internal.platform.ServiceLoaderWrapper;
@@ -115,6 +116,16 @@ public final class PlatformTestStorageRegistry {
     @Override
     public OutputStream openInternalOutputFile(String pathname) throws IOException {
       return new NullOutputStream();
+    }
+
+    @Override
+    public Uri getInputFileUri2(@NonNull String pathname) {
+      return null;
+    }
+
+    @Override
+    public Uri getOutputFileUri2(@NonNull String pathname) {
+      return null;
     }
 
     @Override