+++ services/storage/java/androidx/test/services/storage/TestStorage.java
@@ -77,6 +77,15 @@ public final class TestStorage implements PlatformTestStorage {
   }
 
   /**
+   * @deprecated call getInputFileUri2 instead
+   */
+  @Deprecated
+  public static Uri getInputFileUri(@NonNull String pathname) {
+    checkNotNull(pathname);
+    return HostedFile.buildUri(HostedFile.FileHost.TEST_FILE, pathname);
+  }
+
+  /**
    * Provides a Uri to a test file dependency.
    *
    * <p>In most of the cases, you would use {@link #openInputFile(String)} for opening up an
@@ -90,7 +99,8 @@ public final class TestStorage implements PlatformTestStorage {
    *     "/sdcard/test_input_files/path/to/my_input.txt" on device.
    * @return a content Uri to the test file dependency.
    */
-  public static Uri getInputFileUri(@NonNull String pathname) {
+  @Override
+  public Uri getInputFileUri2(@NonNull String pathname) {
     checkNotNull(pathname);
     return HostedFile.buildUri(HostedFile.FileHost.TEST_FILE, pathname);
   }
@@ -108,6 +118,16 @@ public final class TestStorage implements PlatformTestStorage {
    *     "/path/to/my_output.txt", the file will end up at
    *     "/sdcard/test_output_files/path/to/my_output.txt" on device.
    */
+  @Override
+  public Uri getOutputFileUri2(@NonNull String pathname) {
+    checkNotNull(pathname);
+    return HostedFile.buildUri(HostedFile.FileHost.OUTPUT, pathname);
+  }
+
+  /**
+   * @deprecated call getOutputFileUri2 instead
+   */
+  @Deprecated
   public static Uri getOutputFileUri(@NonNull String pathname) {
     checkNotNull(pathname);
     return HostedFile.buildUri(HostedFile.FileHost.OUTPUT, pathname);